% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{BubStab}
\alias{BubStab}
\title{Calculate the average snip-rebuild distance using bubble clustering}
\usage{
BubStab(data, level, Brad)
}
\arguments{
\item{data}{A numerical matrix}

\item{level}{Level of desired bubble coverage generates n*10^level bubbles are generated}

\item{Brad}{maximum bubble radius in the range of 0-1 of maximum distance between points}
}
\value{
Returns the average snip-rebuild distance for each sample in \code{data} using bubble clustering
}
\description{
Calculate the average snip-rebuild distance using bubble clustering
}
\examples{
BubStab(myData, 4, .75)
BubStab(myData)

}
