% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{BubStabHist}
\alias{BubStabHist}
\title{Calculate the snip-rebuild distance using bubble clustering for each taxa and returns the values as a vector}
\usage{
BubStabHist(data, level, rad = 1)
}
\arguments{
\item{data}{A numerical matrix}

\item{level}{Level of desired bubble coverage generates n*10^level bubbles are generated}

\item{Brad}{maximum bubble radius in the range of 0-1 of maximum distance between points}
}
\value{
Returns the snip-rebuild distance for each sample in \code{data} using bubble clustering as a vector
}
\description{
Calculate the snip-rebuild distance using bubble clustering for each taxa and returns the values as a vector
}
\examples{
BubStab(myData, 4, .75)
BubStab(myData)

}
